Properly clear folder and file names
I noticed an issue with some of my filenames not showing in the file list on the front-end, see image: https://imgur.com/a/dwOlc5d.  Turns out an "&" ampersand in the filename was breaking the output.  Wrapping $folderName and $file_name on lines 572 and 761 in htmlentites() corrected this issue.